+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to the "gruvbox-material" extension will be documented in th
 - Optimize semantic tokens for python.
 - Optimize semantic tokens for C/C++.
 - Optimize semantic tokens for rust.
+- Remove `regenerateThemes` command.
 
 ## 6.3.9