+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the "gruvbox-material" extension will be documented in this file.
 
-## 6.3.10
+## 6.4.0
 
 - Optimize setting colors.
 - Optimize selection colors.